+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Cleaned up some messy spots in the code.
 ## Fixed
 * `Enter` on the numeric keypad also triggers the rename action.
+* Improve paged content rendering on high-DPI screens.
 
 ## [0.1.0] - 2021-02-27
 ## Added